Bug 1799668 - part 12: Merge test tasks
