+++ taskcluster/android_taskgraph/job.py
@@ -36,6 +36,35 @@ gradlew_schema = Schema({
     Optional("dummy-secrets"): [dummy_secret_schema],
 })
 
+run_commands_schema = Schema({
+    Required("using"): "run-commands",
+    Optional("pre-commands"): [[str]],
+    Required("commands"): [[taskref_or_string]],
+    Required("workdir"): str,
+    Optional("use-caches"): bool,
+    Optional("secrets"): [secret_schema],
+    Optional("dummy-secrets"): [dummy_secret_schema],
+})
+
+
+@run_job_using("docker-worker", "run-commands", schema=run_commands_schema)
+def configure_run_commands_schema(config, job, taskdesc):
+    run = job["run"]
+    pre_commands = run.pop("pre-commands", [])
+    pre_commands += [
+        _generate_dummy_secret_command(secret) for secret in run.pop("dummy-secrets", [])
+    ]
+    pre_commands += [
+        _generate_secret_command(secret) for secret in run.get("secrets", [])
+    ]
+
+    all_commands = pre_commands + run.pop("commands", [])
+
+    run["command"] = _convert_commands_to_string(all_commands)
+    _inject_secrets_scopes(run, taskdesc)
+    _set_run_task_attributes(job)
+    configure_taskdesc_for_run(config, job, taskdesc, job["worker"]["implementation"])
+
 
 @run_job_using("docker-worker", "gradlew", schema=gradlew_schema)
 def configure_gradlew(config, job, taskdesc):
